ggml-blas: refactor backend
